many version changes. a few changes to models.

+++ ESI.NET/Models/Character/Agent.cs
@@ -6,16 +6,16 @@ namespace ESI.NET.Models.Character
     public class Agent
     {
         [JsonProperty("agent_id")]
-        public long AgentId { get; set; }
+        public int AgentId { get; set; }
 
         [JsonProperty("points_per_day")]
-        public decimal PointsPerDay { get; set; }
+        public float PointsPerDay { get; set; }
 
         [JsonProperty("remainder_points")]
-        public decimal RemainderPoints { get; set; }
+        public float RemainderPoints { get; set; }
 
         [JsonProperty("skill_type_id")]
-        public long SkillTypeId { get; set; }
+        public int SkillTypeId { get; set; }
 
         [JsonProperty("started_at")]
         public DateTime StartedAt { get; set; }
